+++ system/modules/MonitoringCompression/languages/de/tl_settings.php
@@ -2,7 +2,7 @@
 
 /**
  * Contao Open Source CMS
- * Copyright (C) 2005-2016 Leo Feyer
+ * Copyright (C) 2005-2017 Leo Feyer
  *
  * Formerly known as TYPOlight Open Source CMS.
  *
@@ -21,7 +21,7 @@
  * Software Foundation website at <http://www.gnu.org/licenses/>.
  *
  * PHP version 5
- * @copyright  Cliff Parnitzky 2016-2016
+ * @copyright  Cliff Parnitzky 2016-2017
  * @author     Cliff Parnitzky
  * @package    MonitoringCompression
  * @license    LGPL
@@ -30,6 +30,7 @@
 /**
  * Fields
  */
-$GLOBALS['TL_LANG']['tl_settings']['monitoringAutoCompressionActive'] = array('Automatische Komprimierung aktivieren', 'Wählen Sie ob die automatische Komprimierung aktiviert sein soll.');
+$GLOBALS['TL_LANG']['tl_settings']['monitoringAutoCompressionActive']              = array('Automatische Komprimierung aktivieren', 'Wählen Sie ob die automatische Komprimierung aktiviert sein soll.');
+$GLOBALS['TL_LANG']['tl_settings']['monitoringCompressionResponseTimeCombination'] = array('Antwortzeiten Zusammenfassung', 'Wählen Sie welcher Wert der Antwortzeiten in das komprimierte Testergebnis übernommen werden soll.');
 
 ?>